refactor SideBar/sidebar.tsx to be a single file in components directory
