--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "react-loading": "^2.0.3"
-  }
-}
